Move EventMessage view
